requirements.txt
boto3==1.4.5
botocore==1.5.92
Django==1.11.2
django-analytical==2.2.2
django-cors-headers==2.1.0
django-robots==3.0
django-storages==1.6.5
djangorestframework==3.6.3
docutils==0.14
futures==3.1.1
jmespath==0.9.3
psycopg2==2.7.1
python-dateutil==2.6.1
pytz==2017.2
s3transfer==0.1.10
six==1.10.0
